reorganize routing
